refactor: modularize core files under 300-line limit and clean up dead code
