+++ aura-core/src/torrent/mod.rs
@@ -1,5 +1,6 @@
 pub mod logic;
 pub mod metadata;
+pub mod v2;
 #[cfg(test)]
 mod tests;
 pub use logic::*;
